requirements.txt
openai
python-dotenv
streamlit
schedule
plyer